% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_function.R
\name{agg}
\alias{agg}
\title{Aggregate Composite Indices}
\usage{
agg(X, var_wts = NULL, agg_method = c("ar", "geom"))
}
\arguments{
\item{X}{The composite index dataset}

\item{var_wts}{The vector of variable weights}

\item{agg_method}{The aggregation method. Either 'ar' for arithmetic aggregation or 'geom' for geometric aggregation. Defaults to 'ar'}
}
\value{
A vector of composite scores
}
\description{
This function takes a matrix representing a composite index dataset, a vector of variable weights, and an aggregation method of arithmetic aggregation or geometric aggregation. The variable weights are applied to the columns and then aggregated according to the specified method. It returns a vector of composite scores. If varaible weights are omitted, then it assumes equal weights between the variables.
}
\examples{
agg(X, agg_method = 'geom')

}
