% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{compositeDCC}
\alias{compositeDCC}
\title{compositeDCC (Dynamic Calibrated Composite) with presets for Temp12k}
\usage{
compositeDCC(
  fTS,
  binvec,
  duration = 3000,
  searchRange = c(0, 7000),
  spread = TRUE,
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE,
  ...
)
}
\arguments{
\item{fTS}{TS object of the timeseries that you want to composite}

\item{binvec}{A vector of bin edges for binning}

\item{duration}{The length of time used to create the standardization interval}

\item{searchRange}{An age range from which duration is sampled within}

\item{spread}{An option to spread values over adjacent empty years (if TRUE, spreadPaleoData() is used by binFun)}

\item{gaussianizeInput}{Optionally gaussianize input before compositing}

\item{alignInterpDirection}{Optionally invert timeseries with negative interpretation_direction}

\item{...}{parameters to pass to stanFun}
}
\value{
a list. The main composite is in the list$composite matrix (needs a better definition here) TODO
}
\description{
compositeDCC (Dynamic Calibrated Composite) with presets for Temp12k
}
