% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{compositeEnsembles}
\alias{compositeEnsembles}
\title{Composite with uncertainty}
\usage{
compositeEnsembles(
  fTS,
  binvec,
  spread = TRUE,
  stanFun = standardizeMeanIteratively,
  ageVar = "age",
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE,
  scope = "climate",
  binFun = sampleEnsembleThenBinTs,
  ...
)
}
\arguments{
\item{fTS}{TS object of the timeseries that you want to composite}

\item{binvec}{A vector of bin edges for binning}

\item{spread}{An option to spread values over adjacent empty years (if TRUE, spreadPaleoData() is used by binFun)}

\item{stanFun}{Function to use for standardization}

\item{ageVar}{Name of the time vector in fTS}

\item{gaussianizeInput}{Optionally gaussianize input before compositing}

\item{alignInterpDirection}{Optionally invert timeseries with negative interpretation_direction}

\item{scope}{interpretation_scope for alignInterpDirection (default = 'climate')}

\item{binFun}{Function to use for binning}

\item{...}{parameters to pass to stanFun}
}
\value{
a list. The main composite is in the list$composite matrix (needs a better definition here) TODO
}
\description{
Composite with uncertainty
}
\seealso{
[compositeEnsembles2()], the maintained ensemble-aware compositing function.
}
