% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{compositeEnsembles2}
\alias{compositeEnsembles2}
\title{Composite with uncertainty}
\usage{
compositeEnsembles2(
  fTS,
  binvec,
  nens = 100,
  spread = TRUE,
  stanFun = standardizeMeanIteratively,
  ageVar = "age",
  uncVar = NA,
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE,
  scope = "climate",
  binFun = sampleEnsembleThenBinTs,
  ...
)
}
\arguments{
\item{fTS}{TS object of the timeseries that you want to composite}

\item{binvec}{A vector of bin edges for binning}

\item{nens}{= The number of iterations to randomize over}

\item{spread}{An option to spread values over adjacent empty years (if TRUE, spreadPaleoData() is used by binFun)}

\item{stanFun}{Function to use for standardization}

\item{ageVar}{Name of the time vector in fTS}

\item{uncVar}{Uncertainty value. Enter a string to identify the variable name in fTS. Use a number to provide a unifomr value to all records. This is used for stanFun = sampleEnsembleThenBinTs but not simpleBinTs.}

\item{gaussianizeInput}{Optionally gaussianize input before compositing}

\item{alignInterpDirection}{Optionally invert timeseries with negative interpretation_direction}

\item{scope}{interpretation_scope for alignInterpDirection (default = 'climate')}

\item{binFun}{Function to use for binning}

\item{...}{parameters to pass to stanFun}
}
\value{
a list. The main composite is in the list$composite matrix (needs a better definition here) TODO
}
\description{
Composite with uncertainty
}
