% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{compositeSCC}
\alias{compositeSCC}
\title{compositeSCC (Standard Calibrated Composite) with presets for Temp12k}
\usage{
compositeSCC(fTS, binvec, interval = c(3000, 5000), ...)
}
\arguments{
\item{fTS}{TS object of the timeseries that you want to composite}

\item{binvec}{A vector of bin edges for binning}

\item{interval}{The ager range used to create the standardization interval}

\item{...}{parameters to pass to stanFun}
}
\value{
a list. The main composite is in the list$composite matrix (needs a better definition here) TODO
}
\description{
compositeSCC (Standard Calibrated Composite) with presets for Temp12k
}
