% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_fncs.R
\name{desired_v_shapley}
\alias{desired_v_shapley}
\title{Plot Desired Importance versus Shapley Effects Before and After Optimization}
\usage{
desired_v_shapley(df, desired_impt, shapley_values, optim_shapley)
}
\arguments{
\item{df}{The composite index dataset}

\item{desired_impt}{The vector of desired variable importances}

\item{shapley_values}{The vector of Shapley Effects that result from the original loading of variable weights}

\item{optim_shapley}{The vector of Shapley Effects that result from the optimized variable weights}
}
\value{
A ggplot object depicting the importances
}
\description{
This function takes in three vectors: the desired variable importances, the Shapley Effects resulting from the original variable weight loadings, and the Shapley Effects resulting from the optimized variable weights, and graphs them. It takes the column names from the dataframe df.
}
