% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optim_wts.R
\name{get_optim_wts}
\alias{get_optim_wts}
\title{Estimate new Weights for optimum Shapley Effects}
\usage{
get_optim_wts(
  data,
  model,
  impt,
  randperm = T,
  n.perm = 30,
  n.knn = 10,
  cl = cl,
  ...
)
}
\arguments{
\item{data}{The Composite index}

\item{model}{A function to apply variable weights and aggregate variables to output composite index scores}

\item{impt}{The target vector of variable importances.}

\item{randperm}{Logical parameter to indicate whether parameters}

\item{n.perm}{Number of permutations to sample}

\item{n.knn}{Number of nearest neighbors}

\item{cl}{The cores to be used for the parallel processing}

\item{...}{Additional arguements to pass into the sobolshap_knn function}
}
\value{
A numeric which is the Euclidean distance between estimated Shapley Effects and desired importances.
}
\description{
This function returns the tuned variable weights to achieve the lowest possible difference between the desired Shapley Effects and the estimated Shapley Effects
}
\examples{
cl <- makeCluster(detectCores()-1) # set the number of processor cores
setDefaultCluster(cl=cl)
clusterExport(cl = cl, varlist = list('data', 'model', 'sobolshap_knn', 'weights_shapley_diff'), envir = environment())

get_optim_wts(fn = weights_shapley_diff, lower = rep(0,ncol(data)), upper = rep(4, ncol(data)), control = list(cluster = cl),
        impt = impt, model = model, data = data, method = 'knn', randperm = T, n.perm=30, n.knn=10, agg_method = 'ar')
setDefaultCluster(cl=NULL); stopCluster(cl)

}
