% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{new_composite}
\alias{new_composite}
\title{Make S3 class for compositeEnsembles2}
\usage{
new_composite(x = list())
}
\arguments{
\item{x}{list output from compositeEnsembles2()}
}
\value{
list output from compositeEnsembles2() as S3 class "composite"
}
\description{
Make S3 class for compositeEnsembles2
}
