% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.paleoComposite}
\alias{plot.paleoComposite}
\title{Plot an compoite ensemble}
\usage{
\method{plot}{paleoComposite}(x, ...)
}
\arguments{
\item{x}{The output of compositeEnsembles2()}

\item{...}{additional parameters (see plotComposite)}
}
\value{
a ggplot2 object
}
\description{
Plot an compoite ensemble
}
