% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotComposite}
\alias{plotComposite}
\title{A function to plot the output of compositeEnsembles2}
\usage{
plotComposite(
  compositeList,
  ageUnits = "yr BP",
  valUnits = "standardized",
  title = "composite proxy record",
  combine = TRUE,
  ...
)
}
\arguments{
\item{compositeList}{The output of compositeEnsembles2()}

\item{ageUnits}{units for the age variable}

\item{valUnits}{units for the composite (default = standardized)}

\item{title}{a title for the combined plot (not used if combine == FALSE)}

\item{combine}{output a combined figure using egg::ggarrange or provide a list of subplots}

\item{...}{
  Arguments passed on to \code{\link[lipdViz:plotTimeseriesEnsRibbons]{lipdViz::plotTimeseriesEnsRibbons}}
  \describe{
    \item{\code{probs}}{a vector of probabilities to plot as ribbons. It will create bands as ribbons of quantiles moving inward. If there's an odd number, it plots the middle quantile as a line.}
    \item{\code{color.low}}{Color of the outermost band; the extreme quantiles of the distribution}
    \item{\code{color.high}}{Color of the innermost band; the central quantiles of the distribution}
    \item{\code{color.line}}{Line color (following ggplot rules)}
    \item{\code{color.vector}}{A vector (of length equal to the number of bands) that specifies the colors for the ribbons from the outermost band in (default = NA). Colors specified as string according to ggplot2 conventions. If present, this overrules color.high and color.low}
    \item{\code{line.width}}{Width of the line}
  }}
}
\value{
a combined or list of gg plots
}
\description{
A function to plot the output of compositeEnsembles2
}
