% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.paleoComposite}
\alias{print.paleoComposite}
\title{Print composite output}
\usage{
\method{print}{paleoComposite}(x, ...)
}
\arguments{
\item{x}{composite output}

\item{...}{additional inputs (see printShift)}
}
\description{
Print composite output
}
