% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{printComposite}
\alias{printComposite}
\title{Print compositeEnsembles2() output}
\usage{
printComposite(compositeList)
}
\arguments{
\item{compositeList}{The output of compositeEnsembles2()}
}
\description{
Print compositeEnsembles2() output
}
