% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{rankings}
\alias{rankings}
\title{Non-Compensatory Ranking of Composite Index Observations}
\usage{
rankings(X, method = c("borda", "condorcet"))
}
\arguments{
\item{X}{The composite index dataset}

\item{method}{A voting method for ranked choices c('borda', 'condorcet'). Defaults to 'borda'.}
}
\value{
A votesys object which are the ranked choices.
}
\description{
This function takes in a dataframe which is the composite index dataset (after standardization and variable weights are applied) and ranks observations according to the Borda Count method or the Condorcet method. These methods are non-compensatory aggregation approaches to produce ranks for when a singular composite score is less preferable to a scoreboard- eg. when variables are negatively correlated.
}
