% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{recordRMSE}
\alias{recordRMSE}
\title{recordRMSE}
\usage{
recordRMSE(td, palMat)
}
\arguments{
\item{td}{output of standardizeOverRandomInterval}

\item{palMat}{matrix of paleoData values}
}
\value{
list with RMSE and Bias vectores
}
\description{
recordRMSE
}
