% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{removeConsecutiveDuplicates}
\alias{removeConsecutiveDuplicates}
\title{removeConsecutiveDuplicates}
\usage{
removeConsecutiveDuplicates(spreadValues)
}
\arguments{
\item{spreadValues}{a vector of binned paleoData_values}
}
\value{
a matrix vector
}
\description{
Remove consecutive duplicates within a time series.
}
\examples{
removeConsecutiveDuplicates(c(1, 1, 2, 3, 3, 3, 4, 4, 5, 5))


}
