% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{simpleBinTs}
\alias{simpleBinTs}
\title{simple binning of a lipd_ts object paleoData_values using a nearest neighbor approach}
\usage{
simpleBinTs(
  ts,
  binvec,
  ageVar = "age",
  spread = TRUE,
  spreadBy = abs(mean(diff(binvec)))/10,
  spreadMax = as.numeric(abs(stats::quantile(abs(diff(ts[[ageVar]])), probs = 0.75, na.rm
    = TRUE))),
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE,
  scope = "climate"
)
}
\arguments{
\item{ts}{a lipd_ts object}

\item{binvec}{vector of time boundaries over which to bin}

\item{ageVar}{specify the name the time variable (typically 'age' or 'year')}

\item{spread}{should values be interpolated between bins? (TRUE/FALSE)}

\item{spreadBy}{what is the desired resolution of spread data (passed to spreadPaleoData())}

\item{spreadMax}{a limit to how many years  (given in years) a value can be interpolated across (passed to spreadPaleoData())}

\item{gaussianizeInput}{Force values to gaussian distribution before analysis (TRUE/FALSE)}

\item{alignInterpDirection}{multiply values by -1 if scope_interpDirection == negative (TRUE/FALSE)}

\item{scope}{the scope of the project (typically "climate" (default) or "isotope")}
}
\value{
numeric vector of values of equal length to binvec
}
\description{
simple binning of a lipd_ts object paleoData_values using a nearest neighbor approach
}
