% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{spreadPaleoData}
\alias{spreadPaleoData}
\title{spreadPaleoData}
\usage{
spreadPaleoData(
  age,
  value,
  newAge = NA,
  spreadBy,
  maxGap = NA,
  maxPct = 0.75,
  minAge = -69
)
}
\arguments{
\item{newAge}{new age vector to spread over}

\item{spreadBy}{if newAge is not supplied, what is the desired resolution of spread data.}

\item{minAge}{}
}
\description{
spreadPaleoData
}
