% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{standardize}
\alias{standardize}
\title{Standardize Composite Index Variables}
\usage{
standardize(
  X,
  method = c("min_max", "z_scores", "ratio_mean", "ratio_leader", "distance_mean",
    "distance_leader")
)
}
\arguments{
\item{X}{The composite index dataset.}

\item{method}{The method of standardization. Takes values c('min_max', 'z_scores', 'ratio_mean', 'ratio_leader', 'distance_mean', 'distance_leader') with the default being 'min-max'.}
}
\value{
A matrix with standardized column values.
}
\description{
This function takes a matrix representing a composite index dataset and a method to standardize the columns by. Many common transformations are featured and the default method is to do a min-max transformation..
}
\examples{
standardize(X, agg_method = 'z_scores')

}
