% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardizeMeanIteratively}
\alias{standardizeMeanIteratively}
\title{standardizeMeanIteratively}
\usage{
standardizeMeanIteratively(
  ages,
  pdm,
  duration,
  searchRange,
  normalizeVariance = TRUE,
  thresh = 0.01,
  minN = 8
)
}
\arguments{
\item{ages}{vector of ages}

\item{pdm}{matrix of binned paleodata}

\item{duration}{length of interval over which to standardize}

\item{minN}{}
}
\value{

}
\description{
standardizeMeanIteratively
}
