% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardizeOverInterval}
\alias{standardizeOverInterval}
\title{standardizeOverInterval}
\usage{
standardizeOverInterval(
  ages,
  pdm,
  interval,
  minN = 8,
  normalizeVariance = TRUE
)
}
\arguments{
\item{normalizeVariance}{}
}
\value{

}
\description{
standardizeOverInterval
}
