% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardizeOverRandomInterval}
\alias{standardizeOverRandomInterval}
\title{standardizeOverRandomInterval}
\usage{
standardizeOverRandomInterval(
  ages,
  pdm,
  duration,
  searchRange,
  normalizeVariance = TRUE,
  minN = 8
)
}
\arguments{
\item{minN}{}
}
\value{

}
\description{
standardizeOverRandomInterval
}
