% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Temp12k}
\alias{Temp12k}
\title{Temp12k dataset}
\format{
## `Temp12k`
A list of lipd timeseries
\describe{
  ...
}
}
\source{
<https://lipdverse.org/project/temp12k/>
}
\usage{
Temp12k
}
\description{
The proxy records compiled within for Temp12k dataset
}
\keyword{datasets}
