% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_shapley_diff.R
\name{weights_shapley_diff}
\alias{weights_shapley_diff}
\title{Objective Function to Optimize Over}
\usage{
weights_shapley_diff(
  wts = NULL,
  impt = NULL,
  model = NULL,
  data,
  id.cat = NULL,
  U = NULL,
  method = "knn",
  n.knn = 2,
  return.shap = FALSE,
  randperm = FALSE,
  n.perm = 10000,
  rescale = FALSE,
  n.limit = 2000,
  noise = FALSE,
  ...
)
}
\arguments{
\item{wts}{An initial set of variable weights. If omitted, equal weights are assumed.}

\item{impt}{The target vector of variable importances.}

\item{model}{A function to apply variable weights and aggregate variables to output composite index scores}

\item{data}{The composite index dataset.}

\item{...}{Additional arguements to pass into the sobolshap_knn function}
}
\value{
A numeric which is the Euclidean distance between estimated Shapley Effects and desired importances.
}
\description{
This function is the objective function to be optimized over. It first calculates the Shapley Effects of a set variables in a composite index given a set of variable weights and then calculates the Euclidean distance between the Shapley Effects and a given vector of desired importances. The Shapley Effects are estimated using the sobolshap_knn funciton from the sensitivity package in R. The model is function which applies the variable weights and outputs the composite index scores.
}
\examples{
standardize(X, agg_method = 'z_scores')

}
