% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_fncs.R
\name{wts_v_optim_wts}
\alias{wts_v_optim_wts}
\title{Plot Original Variable Weights versus Optimized Variable Weights}
\usage{
wts_v_optim_wts(df, old_weights, optim_wts)
}
\arguments{
\item{df}{The composite index dataset}

\item{old_weights}{The vector of original variable weights}

\item{optim_wts}{The vector of optimized variable weights}
}
\value{
A ggplot object depicting the weights
}
\description{
This function takes in two vectors: the original variable weights and the optimized variable weights and plots them against one another.
It takes the column names from the dataset df.
}
