% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{compositeEnsembles2}
\alias{compositeEnsembles2}
\title{create paleoclimate composite ensembles from a list of LiPD timeseries. A revised version of compositeEnsembles by Chris Hancock (2024)}
\usage{
compositeEnsembles2(
  fTS,
  binvec,
  nens = 100,
  stanFun = standardizeOverRandomInterval,
  ageVar = "age",
  uncVar = NA,
  weights = NA,
  samplePct = 1,
  scale = FALSE,
  spread = TRUE,
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE,
  scope = "climate",
  binFun = sampleEnsembleThenBinTs,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fTS}{TS list of the timeseries that you want to composite}

\item{binvec}{A vector that describes the edges of the bins to bin to}

\item{nens}{the number of ensembles to create}

\item{stanFun}{function to use for standardization (either standardizeOverRandomInterval, standardizeMeanIteratively, or standardizeOverInterval)}

\item{ageVar}{age variable to use}

\item{uncVar}{uncertainty variable  \cr - Use a string to identify a fTS variable name or use a number to provide a uniform value to all records.  \cr - This is used if stanFun == sampleEnsembleThenBinTs but not for simpleBinTs.}

\item{weights}{weights for calculating the composite mean. \cr - For example, you may want to weight each record in a region relative to the number of total records from a single site.  \cr - Use a string to identify a fTS variable name, otherwise all values will be weighted equally.}

\item{samplePct}{the percent of records to used for each ensemble.  \cr - If length(fTS)*samplePct <= 3, this is ignored.  \cr - Default = 1 which means that 100 percent of records are used for each ensemble}

\item{scale}{scale each composite ensemble member to have a mean of zero and unit variance at the end of the compiting}

\item{spread}{"Spread" data before binning to prevent aliasing of bins (default = TRUE)}

\item{gaussianizeInput}{"Gaussianize" the data before compositing to avoid impacts of irregularly distributed data (default = TRUE)}

\item{alignInterpDirection}{align data by interpretation direction}

\item{scope}{which scope of interpretation to use}

\item{binFun}{function to use for binning (either sampleEnsembleThenBinTs or simpleBinTS)}

\item{verbose}{should a progress bar be printed?}

\item{...}{additional options to pass to stanFun}
}
\value{
a composite list class with three tibbles:
\cr (1) The composite [dims = length(binvec) x (nens + 1 for age column)]
\cr (2) the standardized values [dims = length(binvec) x (length(fTS) + 1 for age column)], and
\cr (3) a count of when proxy values are used in the compilation [same dims as 2]
}
\description{
create paleoclimate composite ensembles from a list of LiPD timeseries. A revised version of compositeEnsembles by Chris Hancock (2024)
}
