% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{plotComposite}
\alias{plotComposite}
\title{Plot compositeEnsembles output}
\usage{
plotComposite(
  compositeList,
  ageUnits = "yr BP",
  valUnits = "standardized",
  title = "composite proxy record",
  combine = TRUE,
  ...
)
}
\arguments{
\item{compositeList}{The output of compositeEnsembles()}

\item{ageUnits}{units for the age variable}

\item{valUnits}{units for the composite}

\item{title}{a title for the combined plot (not used if combine == FALSE)}

\item{combine}{output a combined figure using egg::ggarrange or provide a list of subplots}

\item{...}{parameters to pass to geoChronR::plotTimeseriesEnsRibbons}
}
\value{
a combined or list of gg plots
}
\description{
Plot compositeEnsembles output
}
