% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{removeConsecutiveDuplicates}
\alias{removeConsecutiveDuplicates}
\title{remove consecutive duplicates}
\usage{
removeConsecutiveDuplicates(spreadValues)
}
\arguments{
\item{spreadValues}{}
}
\value{

}
\description{
remove consecutive duplicates
}
