% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{sampleEnsembleThenBinTs}
\alias{sampleEnsembleThenBinTs}
\title{sampleEnsembleThenBinTs}
\usage{
sampleEnsembleThenBinTs(
  ts,
  binvec,
  ageVar = "age",
  uncVar = "paleoData_uncertainty1sd",
  defaultUnc = 1.5,
  ar = sqrt(0.5),
  bamModel = list(ns = 1, name = "bernoulli", param = 0.05),
  spread = TRUE,
  spreadBy = abs(mean(diff(binvec)))/10,
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE,
  scope = "climate"
)
}
\arguments{
\item{scope}{}
}
\value{

}
\description{
sampleEnsembleThenBinTs
}
