% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{simpleBinTs}
\alias{simpleBinTs}
\title{simple binning of a TS object instance}
\usage{
simpleBinTs(
  ts,
  binvec,
  ageVar = "age",
  spread = TRUE,
  spreadBy = abs(mean(diff(binvec)))/10,
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE
)
}
\arguments{
\item{ts}{a ts object}

\item{binvec}{vector of boundaries}

\item{ageVar}{age variable}
}
\value{

}
\description{
simple binning of a TS object instance
}
