% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{simulateAutoCorrelatedUncertainty}
\alias{simulateAutoCorrelatedUncertainty}
\title{simulateAutoCorrelatedUncertainty}
\usage{
simulateAutoCorrelatedUncertainty(
  sd,
  n,
  mean = 0,
  ar = sqrt(0.5),
  arima.order = c(1, 0, 0)
)
}
\arguments{
\item{arima.order}{}
}
\value{

}
\description{
simulateAutoCorrelatedUncertainty
}
