% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{simulateAutoCorrelatedUncertainty}
\alias{simulateAutoCorrelatedUncertainty}
\title{simulateAutoCorrelatedUncertainty}
\usage{
simulateAutoCorrelatedUncertainty(
  sd,
  n,
  mean = 0,
  ar = sqrt(0.5),
  arima.order = c(1, 0, 0)
)
}
\arguments{
\item{sd}{standard deviation of the output vector}

\item{n}{length of output vector}

\item{mean}{mean of the output vector}

\item{ar}{Autocorrelation coefficient to use for modelling uncertainty, what fraction of the uncertainties are autocorrelated? (default = sqrt(0.5); or 50 percent autocorrelated uncertainty)}

\item{arima.order}{Order to use for ARIMA model used in modelling uncertainty (default = c(1,0,0))}
}
\value{
ts simulated from a from an ARIMA model with a defined mean and variance
}
\description{
generate synthetic timeseries
}
