% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardizeOverInterval}
\alias{standardizeOverInterval}
\title{standardizeOverInterval}
\usage{
standardizeOverInterval(
  ages,
  pdm,
  interval,
  minN = 8,
  normalizeVariance = TRUE
)
}
\arguments{
\item{ages}{vector of ages. Each age corresponds to a row in pdm}

\item{pdm}{matrix of binned paleoData values}

\item{interval}{the age range to use for standardizing the data}

\item{minN}{minimum number length of de within the interval to calculate}

\item{normalizeVariance}{Should the pdm be scaled to a uniform mean (FALSE) or variance and mean (TRUE) (default = TRUE)}
}
\value{
a matrix of standardized values
}
\description{
standardizeOverInterval
}
