% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardizeOverRandomInterval}
\alias{standardizeOverRandomInterval}
\title{standardizeOverRandomInterval}
\usage{
standardizeOverRandomInterval(
  ages,
  pdm,
  duration,
  searchRange,
  normalizeVariance = TRUE,
  minN = 8
)
}
\arguments{
\item{ages}{vector of ages. Each age corresponds to a row in pdm}

\item{pdm}{matrix of binned paleoData values}

\item{duration}{length of the interval to standardize the data within}

\item{searchRange}{the age range to which the interval range will be sampled from}

\item{normalizeVariance}{Should the pdm be scaled to a uniform mean (FALSE) or variance and mean (TRUE) (default = TRUE)}

\item{minN}{minimum number length of de within the interval to calculate}
}
\value{
a matrix of standardized values
}
\description{
standardizeOverRandomInterval
}
